Amelioration

+++ adopet-api/routes/dogRoutes.js
@@ -1,8 +1,9 @@
 const express = require('express');
 const Dog = require('../models/dog');
 const router = express.Router();
+const mongoose = require('mongoose');  // Importation de mongoose
 
-// CREATE a new dog
+
 router.post('/dogs', async (req, res) => {
   const { name, age, gender, color, weight, distance, imageUrl, description, owner } = req.body;
 
@@ -25,7 +26,7 @@ router.get('/dogs', async (req, res) => {
   }
 });
 
-// UPDATE a dog by ID
+
 router.put('/dogs/:id', async (req, res) => {
   const { id } = req.params;
   const { name, age, gender, color, weight, distance, imageUrl, description, owner } = req.body;
@@ -39,17 +40,25 @@ router.put('/dogs/:id', async (req, res) => {
   }
 });
 
+
 // DELETE a dog by ID
 router.delete('/dogs/:id', async (req, res) => {
   const { id } = req.params;
 
+  // Check if the ID is valid before attempting to convert it to ObjectId
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ message: 'Invalid Dog ID format' });
+  }
+
   try {
+    // Attempt to delete the dog by ID
     const deletedDog = await Dog.findByIdAndDelete(id);
-    if (!deletedDog) return res.status(404).json({ message: 'Dog not found' });
+    if (!deletedDog) {
+      return res.status(404).json({ message: 'Dog not found' });
+    }
     res.status(200).json({ message: 'Dog deleted successfully' });
   } catch (error) {
     res.status(400).json({ error: error.message });
   }
 });
-
 module.exports = router;
